Move modules into edgepi_rpc_client.
