+++ edgepi_rpc_client/error/rpc_failure.py
