+++ edgepi_rpc_client/services/din/din_pb_enums.py
